refactor: segregate x11 hashing
