+++ src/bench/crypto_hash.cpp
@@ -13,6 +13,7 @@
 #include <crypto/sha512.h>
 #include <crypto/siphash.h>
 #include <hash.h>
+#include <hash_x11.h>
 #include <random.h>
 #include <uint256.h>
