+++ src/crypto/x11/sph_blake.h
@@ -1,8 +1,9 @@
-/* $Id: sph_simd.h 154 2010-04-26 17:00:24Z tp $ */
+/* $Id: sph_blake.h 252 2011-06-07 17:55:14Z tp $ */
 /**
- * SIMD interface. SIMD is a family of functions which differ by
- * their output size; this implementation defines SIMD for output
- * sizes 224, 256, 384 and 512 bits.
+ * BLAKE interface. BLAKE is a family of functions which differ by their
+ * output size; this implementation defines BLAKE for output sizes 224,
+ * 256, 384 and 512 bits. This implementation conforms to the "third
+ * round" specification.
  *
  * ==========================(LICENSE BEGIN)============================
  *
@@ -29,128 +30,140 @@
  *
  * ===========================(LICENSE END)=============================
  *
- * @file     sph_simd.h
+ * @file     sph_blake.h
  * @author   Thomas Pornin <thomas.pornin@cryptolog.com>
  */
 
-#ifndef SPH_SIMD_H__
-#define SPH_SIMD_H__
+#ifndef SPH_BLAKE_H__
+#define SPH_BLAKE_H__
 
 #ifdef __cplusplus
 extern "C"{
 #endif
 
 #include <stddef.h>
-#include <crypto/sph_types.h>
+#include "sph_types.h"
 
 /**
- * Output size (in bits) for SIMD-224.
+ * Output size (in bits) for BLAKE-224.
  */
-#define SPH_SIZE_simd224   224
+#define SPH_SIZE_blake224   224
 
 /**
- * Output size (in bits) for SIMD-256.
+ * Output size (in bits) for BLAKE-256.
  */
-#define SPH_SIZE_simd256   256
+#define SPH_SIZE_blake256   256
+
+#if SPH_64
 
 /**
- * Output size (in bits) for SIMD-384.
+ * Output size (in bits) for BLAKE-384.
  */
-#define SPH_SIZE_simd384   384
+#define SPH_SIZE_blake384   384
 
 /**
- * Output size (in bits) for SIMD-512.
+ * Output size (in bits) for BLAKE-512.
  */
-#define SPH_SIZE_simd512   512
+#define SPH_SIZE_blake512   512
+
+#endif
 
 /**
- * This structure is a context for SIMD computations: it contains the
- * intermediate values and some data from the last entered block. Once
- * an SIMD computation has been performed, the context can be reused for
- * another computation. This specific structure is used for SIMD-224
- * and SIMD-256.
+ * This structure is a context for BLAKE-224 and BLAKE-256 computations:
+ * it contains the intermediate values and some data from the last
+ * entered block. Once a BLAKE computation has been performed, the
+ * context can be reused for another computation.
  *
- * The contents of this structure are private. A running SIMD computation
- * can be cloned by copying the context (e.g. with a simple
+ * The contents of this structure are private. A running BLAKE
+ * computation can be cloned by copying the context (e.g. with a simple
  * <code>memcpy()</code>).
  */
 typedef struct {
 #ifndef DOXYGEN_IGNORE
 	unsigned char buf[64];    /* first field, for alignment */
 	size_t ptr;
-	sph_u32 state[16];
-	sph_u32 count_low, count_high;
+	sph_u32 H[8];
+	sph_u32 S[4];
+	sph_u32 T0, T1;
 #endif
-} sph_simd_small_context;
+} sph_blake_small_context;
 
 /**
- * This structure is a context for SIMD computations: it contains the
- * intermediate values and some data from the last entered block. Once
- * an SIMD computation has been performed, the context can be reused for
- * another computation. This specific structure is used for SIMD-384
- * and SIMD-512.
+ * This structure is a context for BLAKE-224 computations. It is
+ * identical to the common <code>sph_blake_small_context</code>.
+ */
+typedef sph_blake_small_context sph_blake224_context;
+
+/**
+ * This structure is a context for BLAKE-256 computations. It is
+ * identical to the common <code>sph_blake_small_context</code>.
+ */
+typedef sph_blake_small_context sph_blake256_context;
+
+#if SPH_64
+
+/**
+ * This structure is a context for BLAKE-384 and BLAKE-512 computations:
+ * it contains the intermediate values and some data from the last
+ * entered block. Once a BLAKE computation has been performed, the
+ * context can be reused for another computation.
  *
- * The contents of this structure are private. A running SIMD computation
- * can be cloned by copying the context (e.g. with a simple
+ * The contents of this structure are private. A running BLAKE
+ * computation can be cloned by copying the context (e.g. with a simple
  * <code>memcpy()</code>).
  */
 typedef struct {
 #ifndef DOXYGEN_IGNORE
 	unsigned char buf[128];    /* first field, for alignment */
 	size_t ptr;
-	sph_u32 state[32];
-	sph_u32 count_low, count_high;
+	sph_u64 H[8];
+	sph_u64 S[4];
+	sph_u64 T0, T1;
 #endif
-} sph_simd_big_context;
+} sph_blake_big_context;
 
 /**
- * Type for a SIMD-224 context (identical to the common "small" context).
+ * This structure is a context for BLAKE-384 computations. It is
+ * identical to the common <code>sph_blake_small_context</code>.
  */
-typedef sph_simd_small_context sph_simd224_context;
+typedef sph_blake_big_context sph_blake384_context;
 
 /**
- * Type for a SIMD-256 context (identical to the common "small" context).
+ * This structure is a context for BLAKE-512 computations. It is
+ * identical to the common <code>sph_blake_small_context</code>.
  */
-typedef sph_simd_small_context sph_simd256_context;
+typedef sph_blake_big_context sph_blake512_context;
+
+#endif
 
 /**
- * Type for a SIMD-384 context (identical to the common "big" context).
- */
-typedef sph_simd_big_context sph_simd384_context;
-
-/**
- * Type for a SIMD-512 context (identical to the common "big" context).
- */
-typedef sph_simd_big_context sph_simd512_context;
-
-/**
- * Initialize an SIMD-224 context. This process performs no memory allocation.
+ * Initialize a BLAKE-224 context. This process performs no memory allocation.
  *
- * @param cc   the SIMD-224 context (pointer to a
- *             <code>sph_simd224_context</code>)
+ * @param cc   the BLAKE-224 context (pointer to a
+ *             <code>sph_blake224_context</code>)
  */
-void sph_simd224_init(void *cc);
+void sph_blake224_init(void *cc);
 
 /**
  * Process some data bytes. It is acceptable that <code>len</code> is zero
  * (in which case this function does nothing).
  *
- * @param cc     the SIMD-224 context
+ * @param cc     the BLAKE-224 context
  * @param data   the input data
  * @param len    the input data length (in bytes)
  */
-void sph_simd224(void *cc, const void *data, size_t len);
+void sph_blake224(void *cc, const void *data, size_t len);
 
 /**
- * Terminate the current SIMD-224 computation and output the result into
+ * Terminate the current BLAKE-224 computation and output the result into
  * the provided buffer. The destination buffer must be wide enough to
  * accomodate the result (28 bytes). The context is automatically
  * reinitialized.
  *
- * @param cc    the SIMD-224 context
+ * @param cc    the BLAKE-224 context
  * @param dst   the destination buffer
  */
-void sph_simd224_close(void *cc, void *dst);
+void sph_blake224_close(void *cc, void *dst);
 
 /**
  * Add a few additional bits (0 to 7) to the current computation, then
@@ -160,42 +173,42 @@ void sph_simd224_close(void *cc, void *dst);
  * numbered 7 downto 8-n (this is the big-endian convention at the byte
  * level). The context is automatically reinitialized.
  *
- * @param cc    the SIMD-224 context
+ * @param cc    the BLAKE-224 context
  * @param ub    the extra bits
  * @param n     the number of extra bits (0 to 7)
  * @param dst   the destination buffer
  */
-void sph_simd224_addbits_and_close(
+void sph_blake224_addbits_and_close(
 	void *cc, unsigned ub, unsigned n, void *dst);
 
 /**
- * Initialize an SIMD-256 context. This process performs no memory allocation.
+ * Initialize a BLAKE-256 context. This process performs no memory allocation.
  *
- * @param cc   the SIMD-256 context (pointer to a
- *             <code>sph_simd256_context</code>)
+ * @param cc   the BLAKE-256 context (pointer to a
+ *             <code>sph_blake256_context</code>)
  */
-void sph_simd256_init(void *cc);
+void sph_blake256_init(void *cc);
 
 /**
  * Process some data bytes. It is acceptable that <code>len</code> is zero
  * (in which case this function does nothing).
  *
- * @param cc     the SIMD-256 context
+ * @param cc     the BLAKE-256 context
  * @param data   the input data
  * @param len    the input data length (in bytes)
  */
-void sph_simd256(void *cc, const void *data, size_t len);
+void sph_blake256(void *cc, const void *data, size_t len);
 
 /**
- * Terminate the current SIMD-256 computation and output the result into
+ * Terminate the current BLAKE-256 computation and output the result into
  * the provided buffer. The destination buffer must be wide enough to
  * accomodate the result (32 bytes). The context is automatically
  * reinitialized.
  *
- * @param cc    the SIMD-256 context
+ * @param cc    the BLAKE-256 context
  * @param dst   the destination buffer
  */
-void sph_simd256_close(void *cc, void *dst);
+void sph_blake256_close(void *cc, void *dst);
 
 /**
  * Add a few additional bits (0 to 7) to the current computation, then
@@ -205,42 +218,44 @@ void sph_simd256_close(void *cc, void *dst);
  * numbered 7 downto 8-n (this is the big-endian convention at the byte
  * level). The context is automatically reinitialized.
  *
- * @param cc    the SIMD-256 context
+ * @param cc    the BLAKE-256 context
  * @param ub    the extra bits
  * @param n     the number of extra bits (0 to 7)
  * @param dst   the destination buffer
  */
-void sph_simd256_addbits_and_close(
+void sph_blake256_addbits_and_close(
 	void *cc, unsigned ub, unsigned n, void *dst);
 
+#if SPH_64
+
 /**
- * Initialize an SIMD-384 context. This process performs no memory allocation.
+ * Initialize a BLAKE-384 context. This process performs no memory allocation.
  *
- * @param cc   the SIMD-384 context (pointer to a
- *             <code>sph_simd384_context</code>)
+ * @param cc   the BLAKE-384 context (pointer to a
+ *             <code>sph_blake384_context</code>)
  */
-void sph_simd384_init(void *cc);
+void sph_blake384_init(void *cc);
 
 /**
  * Process some data bytes. It is acceptable that <code>len</code> is zero
  * (in which case this function does nothing).
  *
- * @param cc     the SIMD-384 context
+ * @param cc     the BLAKE-384 context
  * @param data   the input data
  * @param len    the input data length (in bytes)
  */
-void sph_simd384(void *cc, const void *data, size_t len);
+void sph_blake384(void *cc, const void *data, size_t len);
 
 /**
- * Terminate the current SIMD-384 computation and output the result into
+ * Terminate the current BLAKE-384 computation and output the result into
  * the provided buffer. The destination buffer must be wide enough to
  * accomodate the result (48 bytes). The context is automatically
  * reinitialized.
  *
- * @param cc    the SIMD-384 context
+ * @param cc    the BLAKE-384 context
  * @param dst   the destination buffer
  */
-void sph_simd384_close(void *cc, void *dst);
+void sph_blake384_close(void *cc, void *dst);
 
 /**
  * Add a few additional bits (0 to 7) to the current computation, then
@@ -250,42 +265,42 @@ void sph_simd384_close(void *cc, void *dst);
  * numbered 7 downto 8-n (this is the big-endian convention at the byte
  * level). The context is automatically reinitialized.
  *
- * @param cc    the SIMD-384 context
+ * @param cc    the BLAKE-384 context
  * @param ub    the extra bits
  * @param n     the number of extra bits (0 to 7)
  * @param dst   the destination buffer
  */
-void sph_simd384_addbits_and_close(
+void sph_blake384_addbits_and_close(
 	void *cc, unsigned ub, unsigned n, void *dst);
 
 /**
- * Initialize an SIMD-512 context. This process performs no memory allocation.
+ * Initialize a BLAKE-512 context. This process performs no memory allocation.
  *
- * @param cc   the SIMD-512 context (pointer to a
- *             <code>sph_simd512_context</code>)
+ * @param cc   the BLAKE-512 context (pointer to a
+ *             <code>sph_blake512_context</code>)
  */
-void sph_simd512_init(void *cc);
+void sph_blake512_init(void *cc);
 
 /**
  * Process some data bytes. It is acceptable that <code>len</code> is zero
  * (in which case this function does nothing).
  *
- * @param cc     the SIMD-512 context
+ * @param cc     the BLAKE-512 context
  * @param data   the input data
  * @param len    the input data length (in bytes)
  */
-void sph_simd512(void *cc, const void *data, size_t len);
+void sph_blake512(void *cc, const void *data, size_t len);
 
 /**
- * Terminate the current SIMD-512 computation and output the result into
+ * Terminate the current BLAKE-512 computation and output the result into
  * the provided buffer. The destination buffer must be wide enough to
  * accomodate the result (64 bytes). The context is automatically
  * reinitialized.
  *
- * @param cc    the SIMD-512 context
+ * @param cc    the BLAKE-512 context
  * @param dst   the destination buffer
  */
-void sph_simd512_close(void *cc, void *dst);
+void sph_blake512_close(void *cc, void *dst);
 
 /**
  * Add a few additional bits (0 to 7) to the current computation, then
@@ -295,13 +310,16 @@ void sph_simd512_close(void *cc, void *dst);
  * numbered 7 downto 8-n (this is the big-endian convention at the byte
  * level). The context is automatically reinitialized.
  *
- * @param cc    the SIMD-512 context
+ * @param cc    the BLAKE-512 context
  * @param ub    the extra bits
  * @param n     the number of extra bits (0 to 7)
  * @param dst   the destination buffer
  */
-void sph_simd512_addbits_and_close(
+void sph_blake512_addbits_and_close(
 	void *cc, unsigned ub, unsigned n, void *dst);
+
+#endif
+
 #ifdef __cplusplus
 }
 #endif
