+++ src/crypto/x11/sph_bmw.h
@@ -1,8 +1,8 @@
-/* $Id: sph_keccak.h 216 2010-06-08 09:46:57Z tp $ */
+/* $Id: sph_bmw.h 216 2010-06-08 09:46:57Z tp $ */
 /**
- * Keccak interface. This is the interface for Keccak with the
- * recommended parameters for SHA-3, with output lengths 224, 256,
- * 384 and 512 bits.
+ * BMW interface. BMW (aka "Blue Midnight Wish") is a family of
+ * functions which differ by their output size; this implementation
+ * defines BMW for output sizes 224, 256, 384 and 512 bits.
  *
  * ==========================(LICENSE BEGIN)============================
  *
@@ -29,111 +29,142 @@
  *
  * ===========================(LICENSE END)=============================
  *
- * @file     sph_keccak.h
+ * @file     sph_bmw.h
  * @author   Thomas Pornin <thomas.pornin@cryptolog.com>
  */
 
-#ifndef SPH_KECCAK_H__
-#define SPH_KECCAK_H__
+#ifndef SPH_BMW_H__
+#define SPH_BMW_H__
 
 #ifdef __cplusplus
 extern "C"{
 #endif
 
 #include <stddef.h>
-#include <crypto/sph_types.h>
+#include "sph_types.h"
 
 /**
- * Output size (in bits) for Keccak-224.
+ * Output size (in bits) for BMW-224.
  */
-#define SPH_SIZE_keccak224   224
+#define SPH_SIZE_bmw224   224
 
 /**
- * Output size (in bits) for Keccak-256.
+ * Output size (in bits) for BMW-256.
  */
-#define SPH_SIZE_keccak256   256
+#define SPH_SIZE_bmw256   256
+
+#if SPH_64
 
 /**
- * Output size (in bits) for Keccak-384.
+ * Output size (in bits) for BMW-384.
  */
-#define SPH_SIZE_keccak384   384
+#define SPH_SIZE_bmw384   384
 
 /**
- * Output size (in bits) for Keccak-512.
+ * Output size (in bits) for BMW-512.
  */
-#define SPH_SIZE_keccak512   512
+#define SPH_SIZE_bmw512   512
+
+#endif
 
 /**
- * This structure is a context for Keccak computations: it contains the
- * intermediate values and some data from the last entered block. Once a
- * Keccak computation has been performed, the context can be reused for
- * another computation.
+ * This structure is a context for BMW-224 and BMW-256 computations:
+ * it contains the intermediate values and some data from the last
+ * entered block. Once a BMW computation has been performed, the
+ * context can be reused for another computation.
  *
- * The contents of this structure are private. A running Keccak computation
- * can be cloned by copying the context (e.g. with a simple
+ * The contents of this structure are private. A running BMW
+ * computation can be cloned by copying the context (e.g. with a simple
  * <code>memcpy()</code>).
  */
 typedef struct {
 #ifndef DOXYGEN_IGNORE
-	unsigned char buf[144];    /* first field, for alignment */
-	size_t ptr, lim;
-	union {
+	unsigned char buf[64];    /* first field, for alignment */
+	size_t ptr;
+	sph_u32 H[16];
 #if SPH_64
-		sph_u64 wide[25];
+	sph_u64 bit_count;
+#else
+	sph_u32 bit_count_high, bit_count_low;
 #endif
-		sph_u32 narrow[50];
-	} u;
 #endif
-} sph_keccak_context;
+} sph_bmw_small_context;
 
 /**
- * Type for a Keccak-224 context (identical to the common context).
+ * This structure is a context for BMW-224 computations. It is
+ * identical to the common <code>sph_bmw_small_context</code>.
  */
-typedef sph_keccak_context sph_keccak224_context;
+typedef sph_bmw_small_context sph_bmw224_context;
 
 /**
- * Type for a Keccak-256 context (identical to the common context).
+ * This structure is a context for BMW-256 computations. It is
+ * identical to the common <code>sph_bmw_small_context</code>.
  */
-typedef sph_keccak_context sph_keccak256_context;
+typedef sph_bmw_small_context sph_bmw256_context;
+
+#if SPH_64
 
 /**
- * Type for a Keccak-384 context (identical to the common context).
- */
-typedef sph_keccak_context sph_keccak384_context;
-
-/**
- * Type for a Keccak-512 context (identical to the common context).
- */
-typedef sph_keccak_context sph_keccak512_context;
-
-/**
- * Initialize a Keccak-224 context. This process performs no memory allocation.
+ * This structure is a context for BMW-384 and BMW-512 computations:
+ * it contains the intermediate values and some data from the last
+ * entered block. Once a BMW computation has been performed, the
+ * context can be reused for another computation.
  *
- * @param cc   the Keccak-224 context (pointer to a
- *             <code>sph_keccak224_context</code>)
+ * The contents of this structure are private. A running BMW
+ * computation can be cloned by copying the context (e.g. with a simple
+ * <code>memcpy()</code>).
  */
-void sph_keccak224_init(void *cc);
+typedef struct {
+#ifndef DOXYGEN_IGNORE
+	unsigned char buf[128];    /* first field, for alignment */
+	size_t ptr;
+	sph_u64 H[16];
+	sph_u64 bit_count;
+#endif
+} sph_bmw_big_context;
+
+/**
+ * This structure is a context for BMW-384 computations. It is
+ * identical to the common <code>sph_bmw_small_context</code>.
+ */
+typedef sph_bmw_big_context sph_bmw384_context;
+
+/**
+ * This structure is a context for BMW-512 computations. It is
+ * identical to the common <code>sph_bmw_small_context</code>.
+ */
+typedef sph_bmw_big_context sph_bmw512_context;
+
+#endif
+
+/**
+ * Initialize a BMW-224 context. This process performs no memory allocation.
+ *
+ * @param cc   the BMW-224 context (pointer to a
+ *             <code>sph_bmw224_context</code>)
+ */
+void sph_bmw224_init(void *cc);
 
 /**
  * Process some data bytes. It is acceptable that <code>len</code> is zero
  * (in which case this function does nothing).
  *
- * @param cc     the Keccak-224 context
+ * @param cc     the BMW-224 context
  * @param data   the input data
  * @param len    the input data length (in bytes)
  */
-void sph_keccak224(void *cc, const void *data, size_t len);
+void sph_bmw224(void *cc, const void *data, size_t len);
 
 /**
- * Terminate the current Keccak-224 computation and output the result into
+ * Terminate the current BMW-224 computation and output the result into
  * the provided buffer. The destination buffer must be wide enough to
  * accomodate the result (28 bytes). The context is automatically
  * reinitialized.
  *
- * @param cc    the Keccak-224 context
+ * @param cc    the BMW-224 context
  * @param dst   the destination buffer
  */
-void sph_keccak224_close(void *cc, void *dst);
+void sph_bmw224_close(void *cc, void *dst);
 
 /**
  * Add a few additional bits (0 to 7) to the current computation, then
@@ -143,42 +174,42 @@ void sph_keccak224_close(void *cc, void *dst);
  * numbered 7 downto 8-n (this is the big-endian convention at the byte
  * level). The context is automatically reinitialized.
  *
- * @param cc    the Keccak-224 context
+ * @param cc    the BMW-224 context
  * @param ub    the extra bits
  * @param n     the number of extra bits (0 to 7)
  * @param dst   the destination buffer
  */
-void sph_keccak224_addbits_and_close(
+void sph_bmw224_addbits_and_close(
 	void *cc, unsigned ub, unsigned n, void *dst);
 
 /**
- * Initialize a Keccak-256 context. This process performs no memory allocation.
+ * Initialize a BMW-256 context. This process performs no memory allocation.
  *
- * @param cc   the Keccak-256 context (pointer to a
- *             <code>sph_keccak256_context</code>)
+ * @param cc   the BMW-256 context (pointer to a
+ *             <code>sph_bmw256_context</code>)
  */
-void sph_keccak256_init(void *cc);
+void sph_bmw256_init(void *cc);
 
 /**
  * Process some data bytes. It is acceptable that <code>len</code> is zero
  * (in which case this function does nothing).
  *
- * @param cc     the Keccak-256 context
+ * @param cc     the BMW-256 context
  * @param data   the input data
  * @param len    the input data length (in bytes)
  */
-void sph_keccak256(void *cc, const void *data, size_t len);
+void sph_bmw256(void *cc, const void *data, size_t len);
 
 /**
- * Terminate the current Keccak-256 computation and output the result into
+ * Terminate the current BMW-256 computation and output the result into
  * the provided buffer. The destination buffer must be wide enough to
  * accomodate the result (32 bytes). The context is automatically
  * reinitialized.
  *
- * @param cc    the Keccak-256 context
+ * @param cc    the BMW-256 context
  * @param dst   the destination buffer
  */
-void sph_keccak256_close(void *cc, void *dst);
+void sph_bmw256_close(void *cc, void *dst);
 
 /**
  * Add a few additional bits (0 to 7) to the current computation, then
@@ -188,42 +219,44 @@ void sph_keccak256_close(void *cc, void *dst);
  * numbered 7 downto 8-n (this is the big-endian convention at the byte
  * level). The context is automatically reinitialized.
  *
- * @param cc    the Keccak-256 context
+ * @param cc    the BMW-256 context
  * @param ub    the extra bits
  * @param n     the number of extra bits (0 to 7)
  * @param dst   the destination buffer
  */
-void sph_keccak256_addbits_and_close(
+void sph_bmw256_addbits_and_close(
 	void *cc, unsigned ub, unsigned n, void *dst);
 
+#if SPH_64
+
 /**
- * Initialize a Keccak-384 context. This process performs no memory allocation.
+ * Initialize a BMW-384 context. This process performs no memory allocation.
  *
- * @param cc   the Keccak-384 context (pointer to a
- *             <code>sph_keccak384_context</code>)
+ * @param cc   the BMW-384 context (pointer to a
+ *             <code>sph_bmw384_context</code>)
  */
-void sph_keccak384_init(void *cc);
+void sph_bmw384_init(void *cc);
 
 /**
  * Process some data bytes. It is acceptable that <code>len</code> is zero
  * (in which case this function does nothing).
  *
- * @param cc     the Keccak-384 context
+ * @param cc     the BMW-384 context
  * @param data   the input data
  * @param len    the input data length (in bytes)
  */
-void sph_keccak384(void *cc, const void *data, size_t len);
+void sph_bmw384(void *cc, const void *data, size_t len);
 
 /**
- * Terminate the current Keccak-384 computation and output the result into
+ * Terminate the current BMW-384 computation and output the result into
  * the provided buffer. The destination buffer must be wide enough to
  * accomodate the result (48 bytes). The context is automatically
  * reinitialized.
  *
- * @param cc    the Keccak-384 context
+ * @param cc    the BMW-384 context
  * @param dst   the destination buffer
  */
-void sph_keccak384_close(void *cc, void *dst);
+void sph_bmw384_close(void *cc, void *dst);
 
 /**
  * Add a few additional bits (0 to 7) to the current computation, then
@@ -233,42 +266,42 @@ void sph_keccak384_close(void *cc, void *dst);
  * numbered 7 downto 8-n (this is the big-endian convention at the byte
  * level). The context is automatically reinitialized.
  *
- * @param cc    the Keccak-384 context
+ * @param cc    the BMW-384 context
  * @param ub    the extra bits
  * @param n     the number of extra bits (0 to 7)
  * @param dst   the destination buffer
  */
-void sph_keccak384_addbits_and_close(
+void sph_bmw384_addbits_and_close(
 	void *cc, unsigned ub, unsigned n, void *dst);
 
 /**
- * Initialize a Keccak-512 context. This process performs no memory allocation.
+ * Initialize a BMW-512 context. This process performs no memory allocation.
  *
- * @param cc   the Keccak-512 context (pointer to a
- *             <code>sph_keccak512_context</code>)
+ * @param cc   the BMW-512 context (pointer to a
+ *             <code>sph_bmw512_context</code>)
  */
-void sph_keccak512_init(void *cc);
+void sph_bmw512_init(void *cc);
 
 /**
  * Process some data bytes. It is acceptable that <code>len</code> is zero
  * (in which case this function does nothing).
  *
- * @param cc     the Keccak-512 context
+ * @param cc     the BMW-512 context
  * @param data   the input data
  * @param len    the input data length (in bytes)
  */
-void sph_keccak512(void *cc, const void *data, size_t len);
+void sph_bmw512(void *cc, const void *data, size_t len);
 
 /**
- * Terminate the current Keccak-512 computation and output the result into
+ * Terminate the current BMW-512 computation and output the result into
  * the provided buffer. The destination buffer must be wide enough to
  * accomodate the result (64 bytes). The context is automatically
  * reinitialized.
  *
- * @param cc    the Keccak-512 context
+ * @param cc    the BMW-512 context
  * @param dst   the destination buffer
  */
-void sph_keccak512_close(void *cc, void *dst);
+void sph_bmw512_close(void *cc, void *dst);
 
 /**
  * Add a few additional bits (0 to 7) to the current computation, then
@@ -278,13 +311,15 @@ void sph_keccak512_close(void *cc, void *dst);
  * numbered 7 downto 8-n (this is the big-endian convention at the byte
  * level). The context is automatically reinitialized.
  *
- * @param cc    the Keccak-512 context
+ * @param cc    the BMW-512 context
  * @param ub    the extra bits
  * @param n     the number of extra bits (0 to 7)
  * @param dst   the destination buffer
  */
-void sph_keccak512_addbits_and_close(
+void sph_bmw512_addbits_and_close(
 	void *cc, unsigned ub, unsigned n, void *dst);
+
+#endif
 
 #ifdef __cplusplus
 }
