+++ src/crypto/x11/sph_cubehash.h
@@ -1,10 +1,9 @@
-/* $Id: sph_shavite.h 208 2010-06-02 20:33:00Z tp $ */
+/* $Id: sph_cubehash.h 180 2010-05-08 02:29:25Z tp $ */
 /**
- * SHAvite-3 interface. This code implements SHAvite-3 with the
- * recommended parameters for SHA-3, with outputs of 224, 256, 384 and
- * 512 bits. In the following, we call the function "SHAvite" (without
- * the "-3" suffix), thus "SHAvite-224" is "SHAvite-3 with a 224-bit
- * output".
+ * CubeHash interface. CubeHash is a family of functions which differ by
+ * their output size; this implementation defines CubeHash for output
+ * sizes 224, 256, 384 and 512 bits, with the "standard parameters"
+ * (CubeHash16/32 with the CubeHash specification notations).
  *
  * ==========================(LICENSE BEGIN)============================
  *
@@ -31,130 +30,107 @@
  *
  * ===========================(LICENSE END)=============================
  *
- * @file     sph_shavite.h
+ * @file     sph_cubehash.h
  * @author   Thomas Pornin <thomas.pornin@cryptolog.com>
  */
 
-#ifndef SPH_SHAVITE_H__
-#define SPH_SHAVITE_H__
-
-#include <stddef.h>
-#include <crypto/sph_types.h>
+#ifndef SPH_CUBEHASH_H__
+#define SPH_CUBEHASH_H__
 
 #ifdef __cplusplus
 extern "C"{
 #endif
 
-/**
- * Output size (in bits) for SHAvite-224.
- */
-#define SPH_SIZE_shavite224   224
+#include <stddef.h>
+#include "sph_types.h"
 
 /**
- * Output size (in bits) for SHAvite-256.
+ * Output size (in bits) for CubeHash-224.
  */
-#define SPH_SIZE_shavite256   256
+#define SPH_SIZE_cubehash224   224
 
 /**
- * Output size (in bits) for SHAvite-384.
+ * Output size (in bits) for CubeHash-256.
  */
-#define SPH_SIZE_shavite384   384
+#define SPH_SIZE_cubehash256   256
 
 /**
- * Output size (in bits) for SHAvite-512.
+ * Output size (in bits) for CubeHash-384.
  */
-#define SPH_SIZE_shavite512   512
+#define SPH_SIZE_cubehash384   384
 
 /**
- * This structure is a context for SHAvite-224 and SHAvite-256 computations:
- * it contains the intermediate values and some data from the last
- * entered block. Once a SHAvite computation has been performed, the
- * context can be reused for another computation.
+ * Output size (in bits) for CubeHash-512.
+ */
+#define SPH_SIZE_cubehash512   512
+
+/**
+ * This structure is a context for CubeHash computations: it contains the
+ * intermediate values and some data from the last entered block. Once
+ * a CubeHash computation has been performed, the context can be reused for
+ * another computation.
  *
- * The contents of this structure are private. A running SHAvite
- * computation can be cloned by copying the context (e.g. with a simple
+ * The contents of this structure are private. A running CubeHash computation
+ * can be cloned by copying the context (e.g. with a simple
  * <code>memcpy()</code>).
  */
 typedef struct {
 #ifndef DOXYGEN_IGNORE
-	unsigned char buf[64];    /* first field, for alignment */
+	unsigned char buf[32];    /* first field, for alignment */
 	size_t ptr;
-	sph_u32 h[8];
-	sph_u32 count0, count1;
+	sph_u32 state[32];
 #endif
-} sph_shavite_small_context;
+} sph_cubehash_context;
 
 /**
- * This structure is a context for SHAvite-224 computations. It is
- * identical to the common <code>sph_shavite_small_context</code>.
+ * Type for a CubeHash-224 context (identical to the common context).
  */
-typedef sph_shavite_small_context sph_shavite224_context;
+typedef sph_cubehash_context sph_cubehash224_context;
 
 /**
- * This structure is a context for SHAvite-256 computations. It is
- * identical to the common <code>sph_shavite_small_context</code>.
+ * Type for a CubeHash-256 context (identical to the common context).
  */
-typedef sph_shavite_small_context sph_shavite256_context;
+typedef sph_cubehash_context sph_cubehash256_context;
 
 /**
- * This structure is a context for SHAvite-384 and SHAvite-512 computations:
- * it contains the intermediate values and some data from the last
- * entered block. Once a SHAvite computation has been performed, the
- * context can be reused for another computation.
+ * Type for a CubeHash-384 context (identical to the common context).
+ */
+typedef sph_cubehash_context sph_cubehash384_context;
+
+/**
+ * Type for a CubeHash-512 context (identical to the common context).
+ */
+typedef sph_cubehash_context sph_cubehash512_context;
+
+/**
+ * Initialize a CubeHash-224 context. This process performs no memory
+ * allocation.
  *
- * The contents of this structure are private. A running SHAvite
- * computation can be cloned by copying the context (e.g. with a simple
- * <code>memcpy()</code>).
+ * @param cc   the CubeHash-224 context (pointer to a
+ *             <code>sph_cubehash224_context</code>)
  */
-typedef struct {
-#ifndef DOXYGEN_IGNORE
-	unsigned char buf[128];    /* first field, for alignment */
-	size_t ptr;
-	sph_u32 h[16];
-	sph_u32 count0, count1, count2, count3;
-#endif
-} sph_shavite_big_context;
-
-/**
- * This structure is a context for SHAvite-384 computations. It is
- * identical to the common <code>sph_shavite_small_context</code>.
- */
-typedef sph_shavite_big_context sph_shavite384_context;
-
-/**
- * This structure is a context for SHAvite-512 computations. It is
- * identical to the common <code>sph_shavite_small_context</code>.
- */
-typedef sph_shavite_big_context sph_shavite512_context;
-
-/**
- * Initialize a SHAvite-224 context. This process performs no memory allocation.
- *
- * @param cc   the SHAvite-224 context (pointer to a
- *             <code>sph_shavite224_context</code>)
- */
-void sph_shavite224_init(void *cc);
+void sph_cubehash224_init(void *cc);
 
 /**
  * Process some data bytes. It is acceptable that <code>len</code> is zero
  * (in which case this function does nothing).
  *
- * @param cc     the SHAvite-224 context
+ * @param cc     the CubeHash-224 context
  * @param data   the input data
  * @param len    the input data length (in bytes)
  */
-void sph_shavite224(void *cc, const void *data, size_t len);
+void sph_cubehash224(void *cc, const void *data, size_t len);
 
 /**
- * Terminate the current SHAvite-224 computation and output the result into
+ * Terminate the current CubeHash-224 computation and output the result into
  * the provided buffer. The destination buffer must be wide enough to
  * accomodate the result (28 bytes). The context is automatically
  * reinitialized.
  *
- * @param cc    the SHAvite-224 context
+ * @param cc    the CubeHash-224 context
  * @param dst   the destination buffer
  */
-void sph_shavite224_close(void *cc, void *dst);
+void sph_cubehash224_close(void *cc, void *dst);
 
 /**
  * Add a few additional bits (0 to 7) to the current computation, then
@@ -164,42 +140,43 @@ void sph_shavite224_close(void *cc, void *dst);
  * numbered 7 downto 8-n (this is the big-endian convention at the byte
  * level). The context is automatically reinitialized.
  *
- * @param cc    the SHAvite-224 context
+ * @param cc    the CubeHash-224 context
  * @param ub    the extra bits
  * @param n     the number of extra bits (0 to 7)
  * @param dst   the destination buffer
  */
-void sph_shavite224_addbits_and_close(
+void sph_cubehash224_addbits_and_close(
 	void *cc, unsigned ub, unsigned n, void *dst);
 
 /**
- * Initialize a SHAvite-256 context. This process performs no memory allocation.
+ * Initialize a CubeHash-256 context. This process performs no memory
+ * allocation.
  *
- * @param cc   the SHAvite-256 context (pointer to a
- *             <code>sph_shavite256_context</code>)
+ * @param cc   the CubeHash-256 context (pointer to a
+ *             <code>sph_cubehash256_context</code>)
  */
-void sph_shavite256_init(void *cc);
+void sph_cubehash256_init(void *cc);
 
 /**
  * Process some data bytes. It is acceptable that <code>len</code> is zero
  * (in which case this function does nothing).
  *
- * @param cc     the SHAvite-256 context
+ * @param cc     the CubeHash-256 context
  * @param data   the input data
  * @param len    the input data length (in bytes)
  */
-void sph_shavite256(void *cc, const void *data, size_t len);
+void sph_cubehash256(void *cc, const void *data, size_t len);
 
 /**
- * Terminate the current SHAvite-256 computation and output the result into
+ * Terminate the current CubeHash-256 computation and output the result into
  * the provided buffer. The destination buffer must be wide enough to
  * accomodate the result (32 bytes). The context is automatically
  * reinitialized.
  *
- * @param cc    the SHAvite-256 context
+ * @param cc    the CubeHash-256 context
  * @param dst   the destination buffer
  */
-void sph_shavite256_close(void *cc, void *dst);
+void sph_cubehash256_close(void *cc, void *dst);
 
 /**
  * Add a few additional bits (0 to 7) to the current computation, then
@@ -209,42 +186,43 @@ void sph_shavite256_close(void *cc, void *dst);
  * numbered 7 downto 8-n (this is the big-endian convention at the byte
  * level). The context is automatically reinitialized.
  *
- * @param cc    the SHAvite-256 context
+ * @param cc    the CubeHash-256 context
  * @param ub    the extra bits
  * @param n     the number of extra bits (0 to 7)
  * @param dst   the destination buffer
  */
-void sph_shavite256_addbits_and_close(
+void sph_cubehash256_addbits_and_close(
 	void *cc, unsigned ub, unsigned n, void *dst);
 
 /**
- * Initialize a SHAvite-384 context. This process performs no memory allocation.
+ * Initialize a CubeHash-384 context. This process performs no memory
+ * allocation.
  *
- * @param cc   the SHAvite-384 context (pointer to a
- *             <code>sph_shavite384_context</code>)
+ * @param cc   the CubeHash-384 context (pointer to a
+ *             <code>sph_cubehash384_context</code>)
  */
-void sph_shavite384_init(void *cc);
+void sph_cubehash384_init(void *cc);
 
 /**
  * Process some data bytes. It is acceptable that <code>len</code> is zero
  * (in which case this function does nothing).
  *
- * @param cc     the SHAvite-384 context
+ * @param cc     the CubeHash-384 context
  * @param data   the input data
  * @param len    the input data length (in bytes)
  */
-void sph_shavite384(void *cc, const void *data, size_t len);
+void sph_cubehash384(void *cc, const void *data, size_t len);
 
 /**
- * Terminate the current SHAvite-384 computation and output the result into
+ * Terminate the current CubeHash-384 computation and output the result into
  * the provided buffer. The destination buffer must be wide enough to
  * accomodate the result (48 bytes). The context is automatically
  * reinitialized.
  *
- * @param cc    the SHAvite-384 context
+ * @param cc    the CubeHash-384 context
  * @param dst   the destination buffer
  */
-void sph_shavite384_close(void *cc, void *dst);
+void sph_cubehash384_close(void *cc, void *dst);
 
 /**
  * Add a few additional bits (0 to 7) to the current computation, then
@@ -254,42 +232,43 @@ void sph_shavite384_close(void *cc, void *dst);
  * numbered 7 downto 8-n (this is the big-endian convention at the byte
  * level). The context is automatically reinitialized.
  *
- * @param cc    the SHAvite-384 context
+ * @param cc    the CubeHash-384 context
  * @param ub    the extra bits
  * @param n     the number of extra bits (0 to 7)
  * @param dst   the destination buffer
  */
-void sph_shavite384_addbits_and_close(
+void sph_cubehash384_addbits_and_close(
 	void *cc, unsigned ub, unsigned n, void *dst);
 
 /**
- * Initialize a SHAvite-512 context. This process performs no memory allocation.
+ * Initialize a CubeHash-512 context. This process performs no memory
+ * allocation.
  *
- * @param cc   the SHAvite-512 context (pointer to a
- *             <code>sph_shavite512_context</code>)
+ * @param cc   the CubeHash-512 context (pointer to a
+ *             <code>sph_cubehash512_context</code>)
  */
-void sph_shavite512_init(void *cc);
+void sph_cubehash512_init(void *cc);
 
 /**
  * Process some data bytes. It is acceptable that <code>len</code> is zero
  * (in which case this function does nothing).
  *
- * @param cc     the SHAvite-512 context
+ * @param cc     the CubeHash-512 context
  * @param data   the input data
  * @param len    the input data length (in bytes)
  */
-void sph_shavite512(void *cc, const void *data, size_t len);
+void sph_cubehash512(void *cc, const void *data, size_t len);
 
 /**
- * Terminate the current SHAvite-512 computation and output the result into
+ * Terminate the current CubeHash-512 computation and output the result into
  * the provided buffer. The destination buffer must be wide enough to
  * accomodate the result (64 bytes). The context is automatically
  * reinitialized.
  *
- * @param cc    the SHAvite-512 context
+ * @param cc    the CubeHash-512 context
  * @param dst   the destination buffer
  */
-void sph_shavite512_close(void *cc, void *dst);
+void sph_cubehash512_close(void *cc, void *dst);
 
 /**
  * Add a few additional bits (0 to 7) to the current computation, then
@@ -299,14 +278,13 @@ void sph_shavite512_close(void *cc, void *dst);
  * numbered 7 downto 8-n (this is the big-endian convention at the byte
  * level). The context is automatically reinitialized.
  *
- * @param cc    the SHAvite-512 context
+ * @param cc    the CubeHash-512 context
  * @param ub    the extra bits
  * @param n     the number of extra bits (0 to 7)
  * @param dst   the destination buffer
  */
-void sph_shavite512_addbits_and_close(
+void sph_cubehash512_addbits_and_close(
 	void *cc, unsigned ub, unsigned n, void *dst);
-
 #ifdef __cplusplus
 }
 #endif
