+++ src/crypto/x11/sph_groestl.h
@@ -1,8 +1,7 @@
-/* $Id: sph_jh.h 216 2010-06-08 09:46:57Z tp $ */
+/* $Id: sph_groestl.h 216 2010-06-08 09:46:57Z tp $ */
 /**
- * JH interface. JH is a family of functions which differ by
- * their output size; this implementation defines JH for output
- * sizes 224, 256, 384 and 512 bits.
+ * Groestl interface. This code implements Groestl with the recommended
+ * parameters for SHA-3, with outputs of 224, 256, 384 and 512 bits.
  *
  * ==========================(LICENSE BEGIN)============================
  *
@@ -29,48 +28,48 @@
  *
  * ===========================(LICENSE END)=============================
  *
- * @file     sph_jh.h
+ * @file     sph_groestl.h
  * @author   Thomas Pornin <thomas.pornin@cryptolog.com>
  */
 
-#ifndef SPH_JH_H__
-#define SPH_JH_H__
+#ifndef SPH_GROESTL_H__
+#define SPH_GROESTL_H__
 
 #ifdef __cplusplus
 extern "C"{
 #endif
 
 #include <stddef.h>
-#include <crypto/sph_types.h>
+#include "sph_types.h"
 
 /**
- * Output size (in bits) for JH-224.
+ * Output size (in bits) for Groestl-224.
  */
-#define SPH_SIZE_jh224   224
+#define SPH_SIZE_groestl224   224
 
 /**
- * Output size (in bits) for JH-256.
+ * Output size (in bits) for Groestl-256.
  */
-#define SPH_SIZE_jh256   256
+#define SPH_SIZE_groestl256   256
 
 /**
- * Output size (in bits) for JH-384.
+ * Output size (in bits) for Groestl-384.
  */
-#define SPH_SIZE_jh384   384
+#define SPH_SIZE_groestl384   384
 
 /**
- * Output size (in bits) for JH-512.
+ * Output size (in bits) for Groestl-512.
  */
-#define SPH_SIZE_jh512   512
+#define SPH_SIZE_groestl512   512
 
 /**
- * This structure is a context for JH computations: it contains the
- * intermediate values and some data from the last entered block. Once
- * a JH computation has been performed, the context can be reused for
- * another computation.
+ * This structure is a context for Groestl-224 and Groestl-256 computations:
+ * it contains the intermediate values and some data from the last
+ * entered block. Once a Groestl computation has been performed, the
+ * context can be reused for another computation.
  *
- * The contents of this structure are private. A running JH computation
- * can be cloned by copying the context (e.g. with a simple
+ * The contents of this structure are private. A running Groestl
+ * computation can be cloned by copying the context (e.g. with a simple
  * <code>memcpy()</code>).
  */
 typedef struct {
@@ -79,66 +78,98 @@ typedef struct {
 	size_t ptr;
 	union {
 #if SPH_64
+		sph_u64 wide[8];
+#endif
+		sph_u32 narrow[16];
+	} state;
+#if SPH_64
+	sph_u64 count;
+#else
+	sph_u32 count_high, count_low;
+#endif
+#endif
+} sph_groestl_small_context;
+
+/**
+ * This structure is a context for Groestl-224 computations. It is
+ * identical to the common <code>sph_groestl_small_context</code>.
+ */
+typedef sph_groestl_small_context sph_groestl224_context;
+
+/**
+ * This structure is a context for Groestl-256 computations. It is
+ * identical to the common <code>sph_groestl_small_context</code>.
+ */
+typedef sph_groestl_small_context sph_groestl256_context;
+
+/**
+ * This structure is a context for Groestl-384 and Groestl-512 computations:
+ * it contains the intermediate values and some data from the last
+ * entered block. Once a Groestl computation has been performed, the
+ * context can be reused for another computation.
+ *
+ * The contents of this structure are private. A running Groestl
+ * computation can be cloned by copying the context (e.g. with a simple
+ * <code>memcpy()</code>).
+ */
+typedef struct {
+#ifndef DOXYGEN_IGNORE
+	unsigned char buf[128];    /* first field, for alignment */
+	size_t ptr;
+	union {
+#if SPH_64
 		sph_u64 wide[16];
 #endif
 		sph_u32 narrow[32];
-	} H;
+	} state;
 #if SPH_64
-	sph_u64 block_count;
+	sph_u64 count;
 #else
-	sph_u32 block_count_high, block_count_low;
+	sph_u32 count_high, count_low;
 #endif
 #endif
-} sph_jh_context;
+} sph_groestl_big_context;
 
 /**
- * Type for a JH-224 context (identical to the common context).
+ * This structure is a context for Groestl-384 computations. It is
+ * identical to the common <code>sph_groestl_small_context</code>.
  */
-typedef sph_jh_context sph_jh224_context;
+typedef sph_groestl_big_context sph_groestl384_context;
 
 /**
- * Type for a JH-256 context (identical to the common context).
+ * This structure is a context for Groestl-512 computations. It is
+ * identical to the common <code>sph_groestl_small_context</code>.
  */
-typedef sph_jh_context sph_jh256_context;
+typedef sph_groestl_big_context sph_groestl512_context;
 
 /**
- * Type for a JH-384 context (identical to the common context).
- */
-typedef sph_jh_context sph_jh384_context;
-
-/**
- * Type for a JH-512 context (identical to the common context).
- */
-typedef sph_jh_context sph_jh512_context;
-
-/**
- * Initialize a JH-224 context. This process performs no memory allocation.
+ * Initialize a Groestl-224 context. This process performs no memory allocation.
  *
- * @param cc   the JH-224 context (pointer to a
- *             <code>sph_jh224_context</code>)
+ * @param cc   the Groestl-224 context (pointer to a
+ *             <code>sph_groestl224_context</code>)
  */
-void sph_jh224_init(void *cc);
+void sph_groestl224_init(void *cc);
 
 /**
  * Process some data bytes. It is acceptable that <code>len</code> is zero
  * (in which case this function does nothing).
  *
- * @param cc     the JH-224 context
+ * @param cc     the Groestl-224 context
  * @param data   the input data
  * @param len    the input data length (in bytes)
  */
-void sph_jh224(void *cc, const void *data, size_t len);
+void sph_groestl224(void *cc, const void *data, size_t len);
 
 /**
- * Terminate the current JH-224 computation and output the result into
+ * Terminate the current Groestl-224 computation and output the result into
  * the provided buffer. The destination buffer must be wide enough to
  * accomodate the result (28 bytes). The context is automatically
  * reinitialized.
  *
- * @param cc    the JH-224 context
+ * @param cc    the Groestl-224 context
  * @param dst   the destination buffer
  */
-void sph_jh224_close(void *cc, void *dst);
+void sph_groestl224_close(void *cc, void *dst);
 
 /**
  * Add a few additional bits (0 to 7) to the current computation, then
@@ -148,42 +179,42 @@ void sph_jh224_close(void *cc, void *dst);
  * numbered 7 downto 8-n (this is the big-endian convention at the byte
  * level). The context is automatically reinitialized.
  *
- * @param cc    the JH-224 context
+ * @param cc    the Groestl-224 context
  * @param ub    the extra bits
  * @param n     the number of extra bits (0 to 7)
  * @param dst   the destination buffer
  */
-void sph_jh224_addbits_and_close(
+void sph_groestl224_addbits_and_close(
 	void *cc, unsigned ub, unsigned n, void *dst);
 
 /**
- * Initialize a JH-256 context. This process performs no memory allocation.
+ * Initialize a Groestl-256 context. This process performs no memory allocation.
  *
- * @param cc   the JH-256 context (pointer to a
- *             <code>sph_jh256_context</code>)
+ * @param cc   the Groestl-256 context (pointer to a
+ *             <code>sph_groestl256_context</code>)
  */
-void sph_jh256_init(void *cc);
+void sph_groestl256_init(void *cc);
 
 /**
  * Process some data bytes. It is acceptable that <code>len</code> is zero
  * (in which case this function does nothing).
  *
- * @param cc     the JH-256 context
+ * @param cc     the Groestl-256 context
  * @param data   the input data
  * @param len    the input data length (in bytes)
  */
-void sph_jh256(void *cc, const void *data, size_t len);
+void sph_groestl256(void *cc, const void *data, size_t len);
 
 /**
- * Terminate the current JH-256 computation and output the result into
+ * Terminate the current Groestl-256 computation and output the result into
  * the provided buffer. The destination buffer must be wide enough to
  * accomodate the result (32 bytes). The context is automatically
  * reinitialized.
  *
- * @param cc    the JH-256 context
+ * @param cc    the Groestl-256 context
  * @param dst   the destination buffer
  */
-void sph_jh256_close(void *cc, void *dst);
+void sph_groestl256_close(void *cc, void *dst);
 
 /**
  * Add a few additional bits (0 to 7) to the current computation, then
@@ -193,42 +224,42 @@ void sph_jh256_close(void *cc, void *dst);
  * numbered 7 downto 8-n (this is the big-endian convention at the byte
  * level). The context is automatically reinitialized.
  *
- * @param cc    the JH-256 context
+ * @param cc    the Groestl-256 context
  * @param ub    the extra bits
  * @param n     the number of extra bits (0 to 7)
  * @param dst   the destination buffer
  */
-void sph_jh256_addbits_and_close(
+void sph_groestl256_addbits_and_close(
 	void *cc, unsigned ub, unsigned n, void *dst);
 
 /**
- * Initialize a JH-384 context. This process performs no memory allocation.
+ * Initialize a Groestl-384 context. This process performs no memory allocation.
  *
- * @param cc   the JH-384 context (pointer to a
- *             <code>sph_jh384_context</code>)
+ * @param cc   the Groestl-384 context (pointer to a
+ *             <code>sph_groestl384_context</code>)
  */
-void sph_jh384_init(void *cc);
+void sph_groestl384_init(void *cc);
 
 /**
  * Process some data bytes. It is acceptable that <code>len</code> is zero
  * (in which case this function does nothing).
  *
- * @param cc     the JH-384 context
+ * @param cc     the Groestl-384 context
  * @param data   the input data
  * @param len    the input data length (in bytes)
  */
-void sph_jh384(void *cc, const void *data, size_t len);
+void sph_groestl384(void *cc, const void *data, size_t len);
 
 /**
- * Terminate the current JH-384 computation and output the result into
+ * Terminate the current Groestl-384 computation and output the result into
  * the provided buffer. The destination buffer must be wide enough to
  * accomodate the result (48 bytes). The context is automatically
  * reinitialized.
  *
- * @param cc    the JH-384 context
+ * @param cc    the Groestl-384 context
  * @param dst   the destination buffer
  */
-void sph_jh384_close(void *cc, void *dst);
+void sph_groestl384_close(void *cc, void *dst);
 
 /**
  * Add a few additional bits (0 to 7) to the current computation, then
@@ -238,42 +269,42 @@ void sph_jh384_close(void *cc, void *dst);
  * numbered 7 downto 8-n (this is the big-endian convention at the byte
  * level). The context is automatically reinitialized.
  *
- * @param cc    the JH-384 context
+ * @param cc    the Groestl-384 context
  * @param ub    the extra bits
  * @param n     the number of extra bits (0 to 7)
  * @param dst   the destination buffer
  */
-void sph_jh384_addbits_and_close(
+void sph_groestl384_addbits_and_close(
 	void *cc, unsigned ub, unsigned n, void *dst);
 
 /**
- * Initialize a JH-512 context. This process performs no memory allocation.
+ * Initialize a Groestl-512 context. This process performs no memory allocation.
  *
- * @param cc   the JH-512 context (pointer to a
- *             <code>sph_jh512_context</code>)
+ * @param cc   the Groestl-512 context (pointer to a
+ *             <code>sph_groestl512_context</code>)
  */
-void sph_jh512_init(void *cc);
+void sph_groestl512_init(void *cc);
 
 /**
  * Process some data bytes. It is acceptable that <code>len</code> is zero
  * (in which case this function does nothing).
  *
- * @param cc     the JH-512 context
+ * @param cc     the Groestl-512 context
  * @param data   the input data
  * @param len    the input data length (in bytes)
  */
-void sph_jh512(void *cc, const void *data, size_t len);
+void sph_groestl512(void *cc, const void *data, size_t len);
 
 /**
- * Terminate the current JH-512 computation and output the result into
+ * Terminate the current Groestl-512 computation and output the result into
  * the provided buffer. The destination buffer must be wide enough to
  * accomodate the result (64 bytes). The context is automatically
  * reinitialized.
  *
- * @param cc    the JH-512 context
+ * @param cc    the Groestl-512 context
  * @param dst   the destination buffer
  */
-void sph_jh512_close(void *cc, void *dst);
+void sph_groestl512_close(void *cc, void *dst);
 
 /**
  * Add a few additional bits (0 to 7) to the current computation, then
@@ -283,12 +314,12 @@ void sph_jh512_close(void *cc, void *dst);
  * numbered 7 downto 8-n (this is the big-endian convention at the byte
  * level). The context is automatically reinitialized.
  *
- * @param cc    the JH-512 context
+ * @param cc    the Groestl-512 context
  * @param ub    the extra bits
  * @param n     the number of extra bits (0 to 7)
  * @param dst   the destination buffer
  */
-void sph_jh512_addbits_and_close(
+void sph_groestl512_addbits_and_close(
 	void *cc, unsigned ub, unsigned n, void *dst);
 
 #ifdef __cplusplus
