+++ src/crypto/x11/sph_skein.h
@@ -1,8 +1,13 @@
-/* $Id: sph_bmw.h 216 2010-06-08 09:46:57Z tp $ */
+/* $Id: sph_skein.h 253 2011-06-07 18:33:10Z tp $ */
 /**
- * BMW interface. BMW (aka "Blue Midnight Wish") is a family of
- * functions which differ by their output size; this implementation
- * defines BMW for output sizes 224, 256, 384 and 512 bits.
+ * Skein interface. The Skein specification defines three main
+ * functions, called Skein-256, Skein-512 and Skein-1024, which can be
+ * further parameterized with an output length. For the SHA-3
+ * competition, Skein-512 is used for output sizes of 224, 256, 384 and
+ * 512 bits; this is what this code implements. Thus, we hereafter call
+ * Skein-224, Skein-256, Skein-384 and Skein-512 what the Skein
+ * specification defines as Skein-512-224, Skein-512-256, Skein-512-384
+ * and Skein-512-512, respectively.
  *
  * ==========================(LICENSE BEGIN)============================
  *
@@ -29,142 +34,109 @@
  *
  * ===========================(LICENSE END)=============================
  *
- * @file     sph_bmw.h
+ * @file     sph_skein.h
  * @author   Thomas Pornin <thomas.pornin@cryptolog.com>
  */
 
-#ifndef SPH_BMW_H__
-#define SPH_BMW_H__
+#ifndef SPH_SKEIN_H__
+#define SPH_SKEIN_H__
 
 #ifdef __cplusplus
 extern "C"{
 #endif
 
 #include <stddef.h>
-#include <crypto/sph_types.h>
-
-/**
- * Output size (in bits) for BMW-224.
- */
-#define SPH_SIZE_bmw224   224
-
-/**
- * Output size (in bits) for BMW-256.
- */
-#define SPH_SIZE_bmw256   256
+#include "sph_types.h"
 
 #if SPH_64
 
 /**
- * Output size (in bits) for BMW-384.
+ * Output size (in bits) for Skein-224.
  */
-#define SPH_SIZE_bmw384   384
+#define SPH_SIZE_skein224   224
 
 /**
- * Output size (in bits) for BMW-512.
+ * Output size (in bits) for Skein-256.
  */
-#define SPH_SIZE_bmw512   512
-
-#endif
+#define SPH_SIZE_skein256   256
 
 /**
- * This structure is a context for BMW-224 and BMW-256 computations:
- * it contains the intermediate values and some data from the last
- * entered block. Once a BMW computation has been performed, the
- * context can be reused for another computation.
+ * Output size (in bits) for Skein-384.
+ */
+#define SPH_SIZE_skein384   384
+
+/**
+ * Output size (in bits) for Skein-512.
+ */
+#define SPH_SIZE_skein512   512
+
+/**
+ * This structure is a context for Skein computations (with a 384- or
+ * 512-bit output): it contains the intermediate values and some data
+ * from the last entered block. Once a Skein computation has been
+ * performed, the context can be reused for another computation.
  *
- * The contents of this structure are private. A running BMW
- * computation can be cloned by copying the context (e.g. with a simple
+ * The contents of this structure are private. A running Skein computation
+ * can be cloned by copying the context (e.g. with a simple
  * <code>memcpy()</code>).
  */
 typedef struct {
 #ifndef DOXYGEN_IGNORE
 	unsigned char buf[64];    /* first field, for alignment */
 	size_t ptr;
-	sph_u32 H[16];
-#if SPH_64
-	sph_u64 bit_count;
-#else
-	sph_u32 bit_count_high, bit_count_low;
+	sph_u64 h0, h1, h2, h3, h4, h5, h6, h7;
+	sph_u64 bcount;
 #endif
-#endif
-} sph_bmw_small_context;
+} sph_skein_big_context;
 
 /**
- * This structure is a context for BMW-224 computations. It is
- * identical to the common <code>sph_bmw_small_context</code>.
+ * Type for a Skein-224 context (identical to the common "big" context).
  */
-typedef sph_bmw_small_context sph_bmw224_context;
+typedef sph_skein_big_context sph_skein224_context;
 
 /**
- * This structure is a context for BMW-256 computations. It is
- * identical to the common <code>sph_bmw_small_context</code>.
+ * Type for a Skein-256 context (identical to the common "big" context).
  */
-typedef sph_bmw_small_context sph_bmw256_context;
-
-#if SPH_64
+typedef sph_skein_big_context sph_skein256_context;
 
 /**
- * This structure is a context for BMW-384 and BMW-512 computations:
- * it contains the intermediate values and some data from the last
- * entered block. Once a BMW computation has been performed, the
- * context can be reused for another computation.
+ * Type for a Skein-384 context (identical to the common "big" context).
+ */
+typedef sph_skein_big_context sph_skein384_context;
+
+/**
+ * Type for a Skein-512 context (identical to the common "big" context).
+ */
+typedef sph_skein_big_context sph_skein512_context;
+
+/**
+ * Initialize a Skein-224 context. This process performs no memory allocation.
  *
- * The contents of this structure are private. A running BMW
- * computation can be cloned by copying the context (e.g. with a simple
- * <code>memcpy()</code>).
+ * @param cc   the Skein-224 context (pointer to a
+ *             <code>sph_skein224_context</code>)
  */
-typedef struct {
-#ifndef DOXYGEN_IGNORE
-	unsigned char buf[128];    /* first field, for alignment */
-	size_t ptr;
-	sph_u64 H[16];
-	sph_u64 bit_count;
-#endif
-} sph_bmw_big_context;
-
-/**
- * This structure is a context for BMW-384 computations. It is
- * identical to the common <code>sph_bmw_small_context</code>.
- */
-typedef sph_bmw_big_context sph_bmw384_context;
-
-/**
- * This structure is a context for BMW-512 computations. It is
- * identical to the common <code>sph_bmw_small_context</code>.
- */
-typedef sph_bmw_big_context sph_bmw512_context;
-
-#endif
-
-/**
- * Initialize a BMW-224 context. This process performs no memory allocation.
- *
- * @param cc   the BMW-224 context (pointer to a
- *             <code>sph_bmw224_context</code>)
- */
-void sph_bmw224_init(void *cc);
+void sph_skein224_init(void *cc);
 
 /**
  * Process some data bytes. It is acceptable that <code>len</code> is zero
  * (in which case this function does nothing).
  *
- * @param cc     the BMW-224 context
+ * @param cc     the Skein-224 context
  * @param data   the input data
  * @param len    the input data length (in bytes)
  */
-void sph_bmw224(void *cc, const void *data, size_t len);
+void sph_skein224(void *cc, const void *data, size_t len);
 
 /**
- * Terminate the current BMW-224 computation and output the result into
+ * Terminate the current Skein-224 computation and output the result into
  * the provided buffer. The destination buffer must be wide enough to
  * accomodate the result (28 bytes). The context is automatically
  * reinitialized.
  *
- * @param cc    the BMW-224 context
+ * @param cc    the Skein-224 context
  * @param dst   the destination buffer
  */
-void sph_bmw224_close(void *cc, void *dst);
+void sph_skein224_close(void *cc, void *dst);
 
 /**
  * Add a few additional bits (0 to 7) to the current computation, then
@@ -174,42 +146,42 @@ void sph_bmw224_close(void *cc, void *dst);
  * numbered 7 downto 8-n (this is the big-endian convention at the byte
  * level). The context is automatically reinitialized.
  *
- * @param cc    the BMW-224 context
+ * @param cc    the Skein-224 context
  * @param ub    the extra bits
  * @param n     the number of extra bits (0 to 7)
  * @param dst   the destination buffer
  */
-void sph_bmw224_addbits_and_close(
+void sph_skein224_addbits_and_close(
 	void *cc, unsigned ub, unsigned n, void *dst);
 
 /**
- * Initialize a BMW-256 context. This process performs no memory allocation.
+ * Initialize a Skein-256 context. This process performs no memory allocation.
  *
- * @param cc   the BMW-256 context (pointer to a
- *             <code>sph_bmw256_context</code>)
+ * @param cc   the Skein-256 context (pointer to a
+ *             <code>sph_skein256_context</code>)
  */
-void sph_bmw256_init(void *cc);
+void sph_skein256_init(void *cc);
 
 /**
  * Process some data bytes. It is acceptable that <code>len</code> is zero
  * (in which case this function does nothing).
  *
- * @param cc     the BMW-256 context
+ * @param cc     the Skein-256 context
  * @param data   the input data
  * @param len    the input data length (in bytes)
  */
-void sph_bmw256(void *cc, const void *data, size_t len);
+void sph_skein256(void *cc, const void *data, size_t len);
 
 /**
- * Terminate the current BMW-256 computation and output the result into
+ * Terminate the current Skein-256 computation and output the result into
  * the provided buffer. The destination buffer must be wide enough to
  * accomodate the result (32 bytes). The context is automatically
  * reinitialized.
  *
- * @param cc    the BMW-256 context
+ * @param cc    the Skein-256 context
  * @param dst   the destination buffer
  */
-void sph_bmw256_close(void *cc, void *dst);
+void sph_skein256_close(void *cc, void *dst);
 
 /**
  * Add a few additional bits (0 to 7) to the current computation, then
@@ -219,44 +191,42 @@ void sph_bmw256_close(void *cc, void *dst);
  * numbered 7 downto 8-n (this is the big-endian convention at the byte
  * level). The context is automatically reinitialized.
  *
- * @param cc    the BMW-256 context
+ * @param cc    the Skein-256 context
  * @param ub    the extra bits
  * @param n     the number of extra bits (0 to 7)
  * @param dst   the destination buffer
  */
-void sph_bmw256_addbits_and_close(
+void sph_skein256_addbits_and_close(
 	void *cc, unsigned ub, unsigned n, void *dst);
 
-#if SPH_64
-
 /**
- * Initialize a BMW-384 context. This process performs no memory allocation.
+ * Initialize a Skein-384 context. This process performs no memory allocation.
  *
- * @param cc   the BMW-384 context (pointer to a
- *             <code>sph_bmw384_context</code>)
+ * @param cc   the Skein-384 context (pointer to a
+ *             <code>sph_skein384_context</code>)
  */
-void sph_bmw384_init(void *cc);
+void sph_skein384_init(void *cc);
 
 /**
  * Process some data bytes. It is acceptable that <code>len</code> is zero
  * (in which case this function does nothing).
  *
- * @param cc     the BMW-384 context
+ * @param cc     the Skein-384 context
  * @param data   the input data
  * @param len    the input data length (in bytes)
  */
-void sph_bmw384(void *cc, const void *data, size_t len);
+void sph_skein384(void *cc, const void *data, size_t len);
 
 /**
- * Terminate the current BMW-384 computation and output the result into
+ * Terminate the current Skein-384 computation and output the result into
  * the provided buffer. The destination buffer must be wide enough to
  * accomodate the result (48 bytes). The context is automatically
  * reinitialized.
  *
- * @param cc    the BMW-384 context
+ * @param cc    the Skein-384 context
  * @param dst   the destination buffer
  */
-void sph_bmw384_close(void *cc, void *dst);
+void sph_skein384_close(void *cc, void *dst);
 
 /**
  * Add a few additional bits (0 to 7) to the current computation, then
@@ -266,42 +236,42 @@ void sph_bmw384_close(void *cc, void *dst);
  * numbered 7 downto 8-n (this is the big-endian convention at the byte
  * level). The context is automatically reinitialized.
  *
- * @param cc    the BMW-384 context
+ * @param cc    the Skein-384 context
  * @param ub    the extra bits
  * @param n     the number of extra bits (0 to 7)
  * @param dst   the destination buffer
  */
-void sph_bmw384_addbits_and_close(
+void sph_skein384_addbits_and_close(
 	void *cc, unsigned ub, unsigned n, void *dst);
 
 /**
- * Initialize a BMW-512 context. This process performs no memory allocation.
+ * Initialize a Skein-512 context. This process performs no memory allocation.
  *
- * @param cc   the BMW-512 context (pointer to a
- *             <code>sph_bmw512_context</code>)
+ * @param cc   the Skein-512 context (pointer to a
+ *             <code>sph_skein512_context</code>)
  */
-void sph_bmw512_init(void *cc);
+void sph_skein512_init(void *cc);
 
 /**
  * Process some data bytes. It is acceptable that <code>len</code> is zero
  * (in which case this function does nothing).
  *
- * @param cc     the BMW-512 context
+ * @param cc     the Skein-512 context
  * @param data   the input data
  * @param len    the input data length (in bytes)
  */
-void sph_bmw512(void *cc, const void *data, size_t len);
+void sph_skein512(void *cc, const void *data, size_t len);
 
 /**
- * Terminate the current BMW-512 computation and output the result into
+ * Terminate the current Skein-512 computation and output the result into
  * the provided buffer. The destination buffer must be wide enough to
  * accomodate the result (64 bytes). The context is automatically
  * reinitialized.
  *
- * @param cc    the BMW-512 context
+ * @param cc    the Skein-512 context
  * @param dst   the destination buffer
  */
-void sph_bmw512_close(void *cc, void *dst);
+void sph_skein512_close(void *cc, void *dst);
 
 /**
  * Add a few additional bits (0 to 7) to the current computation, then
@@ -311,12 +281,12 @@ void sph_bmw512_close(void *cc, void *dst);
  * numbered 7 downto 8-n (this is the big-endian convention at the byte
  * level). The context is automatically reinitialized.
  *
- * @param cc    the BMW-512 context
+ * @param cc    the Skein-512 context
  * @param ub    the extra bits
  * @param n     the number of extra bits (0 to 7)
  * @param dst   the destination buffer
  */
-void sph_bmw512_addbits_and_close(
+void sph_skein512_addbits_and_close(
 	void *cc, unsigned ub, unsigned n, void *dst);
 
 #endif
